Use an interval timer instead of nanosleep
The existing code (introduced in 796b661284ea296833424f7f55aff2da3ba08d81) uses
the high-resolution clock to measure the time spent processing the args.  The
value passed to nanosleep is reduced accordingly, attempting to make the time
between samples close to the given interval.

However, it was noticed during testing that the actual sample interval was
mostly greater than the given interval, resulting in sampling drift.

This commit replaces the nanosleep technique with an interval timer to improve
the accuracy of the sample interval.

To test the accuracy of the sample interval, the old slstatus (which uses
clock_gettime and nanosleep) and the new slstatus (which uses an interval timer
and sigsuspend) were run simultaneously for a long time with a given interval of
2000 ms.  The clockdiff component was the 2nd field in config.h for both.

The old slstatus created 33001 samples with a mean sample interval of
2.0055363936547 s (an error of 5.536 ms).

The new slstatus created 33093 samples with a mean sample interval of
2.0000005932249 s (an error of 0.593 µs).

Here are bash commands to reproduce the findings.

$ ./slstatus-sleep -s > slstatus-sleep.txt & \
  ./slstatus-alarm -s > slstatus-alarm.txt &

Let the slstatus commands run for a long time, then kill them simultaneously.

$ datamash --header-out --narm min 2 max 2 range 2 mean 2 iqr 2 pstdev 2 mad 2 madraw 2 pskew 2 pkurt 2 < slstatus-sleep.txt | perl -pe 's/\(.+?-2\)//g' | datamash transpose > slstatus-sleep.stats.txt
$ datamash --header-out --narm min 2 max 2 range 2 mean 2 iqr 2 pstdev 2 mad 2 madraw 2 pskew 2 pkurt 2 < slstatus-alarm.txt | perl -pe 's/\(.+?-2\)//g' | datamash transpose > slstatus-alarm.stats.txt
$ { printf 'stat sleep alarm\n' ; join slstatus-sleep.stats.txt slstatus-alarm.stats.txt ; } | column --table

stat    sleep              alarm
min     1.934483           1.847595
max     2.546218           2.152546
range   0.611735           0.304951
mean    2.0055363936547    2.0000005932249
iqr     0.006044           0.00362425
pstdev  0.013058775331363  0.016657573454672
mad     0.0011297412       0.0026886951
madraw  0.000762           0.0018135
pskew   8.9346271610893    -0.057764332015856
pkurt   171.38822823919    21.36680567135