Add meters for showing values graphically
The functions lower_blocks_1, hor_lines_1, upper_blocks_1 take a real number in the interval [0, 1] and return a wide character from their respective block group, indicating the magnitude of the given number.

The functions left_blocks_meter, ver_lines_meter, right_blocks_meter take a real number in the interval [0, 1], a pointer to wide character array, and the capacity of the array, and fill in the array with wide characters from their respective block group, giving it the appearance of a meter, indicating the magnitude of the given number.

The wide characters used are found here:

  Block Elements
  Range: 2580–259F
  https://www.unicode.org/charts/PDF/U2580.pdf

  Symbols for Legacy Computing
  Range: 1FB00–1FBFF
  https://www.unicode.org/charts/PDF/U1FB00.pdf